# lib/api-zod/src/generated/types/attachmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttachmentInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /** @maxLength 120 */
  type: string;
  /** @maxLength 6000000 */
  content: string;
}