# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  workspaceId: string;
  name: string;
  prompt: string;
  /** @nullable */
  html: string | null;
  createdAt: Date;
  updatedAt: Date;
}