# lib/api-zod/src/generated/types/projectChatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttachmentInput } from './attachmentInput';

export interface ProjectChatInput {
  /**
     * @minLength 8
     * @maxLength 128
     */
  workspaceId: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  message: string;
  /** @maxItems 5 */
  attachments?: AttachmentInput[];
}