# lib/api-zod/src/generated/types/projectChatResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';
import type { ProjectMessage } from './projectMessage';

export interface ProjectChatResult {
  project: Project;
  userMessage: ProjectMessage;
  assistantMessage: ProjectMessage;
}