# lib/api-zod/src/generated/types/projectMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectMessageRole } from './projectMessageRole';

export interface ProjectMessage {
  id: number;
  projectId: number;
  role: ProjectMessageRole;
  content: string;
  createdAt: Date;
}