# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  /**
     * @minLength 8
     * @maxLength 128
     */
  workspaceId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /** @maxLength 10000 */
  prompt?: string;
  /**
     * @maxLength 2000000
     * @nullable
     */
  html?: string | null;
}