# lib/api-zod/src/generated/types/projectWorkspaceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectWorkspaceInput {
  /**
     * @minLength 8
     * @maxLength 128
     */
  workspaceId: string;
}